Decoder is working. Almost.

+++ include/jpeg2d_decoder.h
@@ -1,10 +1,9 @@
 #ifndef JPEG2D_DECODER_H
 #define JPEG2D_DECODER_H
 
-#include "jpeg2d.h"
 #include "huffman_decoder.h"
+#include "jpeg2d.h"
 
-#include <vector>
 #include <string>
 
 class JPEG2DDecoder {
@@ -27,7 +26,7 @@ private:
   void dezigzag();
   void dequantize();
   void backwardDCT();
-  static void backwardDCTBlock(const Block<int8_t> &input, Block<double> &output);
+  static void backwardDCTBlock(const Block<double> &input, Block<uint8_t> &output);
   void deblockize();
   void YCbCrToRGB();
 
@@ -63,13 +62,13 @@ private:
   std::vector<Block<int8_t>> m_channel_Cb_unzigzaged;
   std::vector<Block<int8_t>> m_channel_Cr_unzigzaged;
 
-  std::vector<Block<int8_t>> m_channel_Y_dequantized;
-  std::vector<Block<int8_t>> m_channel_Cb_dequantized;
-  std::vector<Block<int8_t>> m_channel_Cr_dequantized;
+  std::vector<Block<double>> m_channel_Y_dequantized;
+  std::vector<Block<double>> m_channel_Cb_dequantized;
+  std::vector<Block<double>> m_channel_Cr_dequantized;
 
-  std::vector<Block<double>> m_channel_Y_detransformed;
-  std::vector<Block<double>> m_channel_Cb_detransformed;
-  std::vector<Block<double>> m_channel_Cr_detransformed;
+  std::vector<Block<uint8_t>> m_channel_Y_detransformed;
+  std::vector<Block<uint8_t>> m_channel_Cb_detransformed;
+  std::vector<Block<uint8_t>> m_channel_Cr_detransformed;
 
   std::vector<uint8_t> m_channel_Y_deblockized;
   std::vector<uint8_t> m_channel_Cb_deblockized;
